Moved docker writing to a common location.
